persona_bot/requirements.txt
aiogram==3.4.1
openai
python-dotenv
sqlite3